+++ game/shared/voice_status.cpp
@@ -348,11 +348,11 @@ void CVoiceStatus::UpdateServerState(bool bForce)
 	Q_strncpy(str, "vban", sizeof(str));
 	bool bChange = false;
 
-	for(unsigned long dw = 0; dw < VOICE_MAX_PLAYERS_DW; dw++)
+	for(unsigned int dw = 0; dw < VOICE_MAX_PLAYERS_DW; dw++)
 	{
-		unsigned long serverBanMask = 0;
-		unsigned long banMask = 0;
-		for(unsigned long i = 0; i < 32; i++)
+		unsigned int serverBanMask = 0;
+		unsigned int banMask = 0;
+		for(unsigned int i = 0; i < 32; i++)
 		{
 			int playerIndex = (dw * 32 + i);
 			if(playerIndex >= MAX_PLAYERS)
@@ -410,8 +410,8 @@ void CVoiceStatus::HandleVoiceMaskMsg(bf_read &msg)
 	unsigned int dw;
 	for(dw = 0; dw < VOICE_MAX_PLAYERS_DW; dw++)
 	{
-		m_AudiblePlayers.SetDWord(dw, (unsigned long)msg.ReadLong());
-		m_ServerBannedPlayers.SetDWord(dw, (unsigned long)msg.ReadLong());
+		m_AudiblePlayers.SetDWord(dw, (unsigned int)msg.ReadLong());
+		m_ServerBannedPlayers.SetDWord(dw, (unsigned int)msg.ReadLong());
 
 		if(voice_clientdebug.GetInt())
 		{
